initial import of lumberjack.
lumberjack is a syslog to bunyan translator.

it current supports listening for syslog messages on a tcp socket.
cos syslog is crap, the best it can do is parse the syslog level,
facility, and date off the wire. the rest of a syslog message is
pretty arbitrary so it's logged directly as the msg. it's assumed
something will postprocess it later.

it's a bit rough at the moment, but im committing it so i can keep
track of things as theyre changed.